Provide DI for Network
Continuing the exercise, the UsdaApi object is now being provided by
dependency injection to the two Presenter implementations in this
application.

This really showcases one of the advantages of DI, since the duplicated
code for creating the UsdaApi objects was avoided. Also, it was simple
to ensure that the UsdaApi object was a Singleton.

This example shows a nice and minimal 'cascade' of the dependency
injection graph, since UsdaApi needs a Retrofit object to be created,
which then needs a Converter.Factory and a String to be instantiated as
well. The NetworkModule created here handles all the methods that
provide these dependencies.